Exploratory Data Analysis using sql
Exploratory Data Analysis using sql. This is a self Exploratory Data Analysis project that was done in order to demonstrate some Data Analysis Skills. The covid-19 data used was obtained from the site ourworldindata.org. The downloaded dataset was first cleaned using spreadsheet software, then saved to an SQL database in the form of tables. A jupyter notebook environment was used to run the python and sqllite code for the data exploration. The idea behind the project can be applied on any dataset so as to get more insights about the data and to see some interesting trends.